# src/swagger/swagger2_0.json
{
    "swagger": "2.0",
    "info": {
        "title": "MEF Chat API",
        "description": "API endpoints for a chat services documented on swagger",
        "contact": {
            "name": "Dhaval Koradiya",
            "email": "dhaval.koradiya@tcs.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "apiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/api/userdetail": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "properties": {
                                "birthYear": {
                                    "example": null,
                                    "x-nullable": true
                                },
                                "mobileNumber": {
                                    "example": "982398293",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "User"
                ],
                "description": "userdetail",
                "operationId": "userdetail",
                "summary": "userdetail",
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        }
    },
...
}